Updated to R-devel revision 77254.

+++ R/src/main/connections.c
@@ -126,7 +126,7 @@ typedef long long int _lli_t;
 # include <Startup.h>
 #endif
 
-#define NCONNECTIONS 128 /* snow needs one per slave node */
+#define NCONNECTIONS 128 /* snow needs one per no-echo node */
 #define NSINKS 21
 
 static Rconnection Connections[NCONNECTIONS];
@@ -1127,7 +1127,7 @@ static size_t fifo_read(void *ptr, size_t size, size_t nitems,
     Rfifoconn thiscon = con->private;
 
     /* uses 'size_t' for len */
-    if ((double) size * (double) nitems > SSIZE_MAX)
+    if ((double) size * (double) nitems > (double) SSIZE_MAX)
 	error(_("specified block is too large"));
     return read(thiscon->fd, ptr, size * nitems)/size;
 }
@@ -1138,7 +1138,7 @@ static size_t fifo_write(const void *ptr, size_t size, size_t nitems,
     Rfifoconn thiscon = con->private;
 
     /* uses 'size_t' for len */
-    if ((double) size * (double) nitems > SSIZE_MAX)
+    if ((double) size * (double) nitems > (double) SSIZE_MAX)
       error(_("specified block is too large"));
     return write(thiscon->fd, ptr, size * nitems)/size;
 }
@@ -2991,7 +2991,7 @@ static void text_init(Rconnection con, SEXP text, int type)
 	    (double) strlen(type == 1 ? translateChar(STRING_ELT(text, i))
 			    : ((type == 3) ?translateCharUTF8(STRING_ELT(text, i))
 			       : CHAR(STRING_ELT(text, i))) ) + 1;
-    if (dnc >= SIZE_MAX)
+    if (dnc >= (double) SIZE_MAX)
 	error(_("too many characters for text connection"));
     else nchars = (size_t) dnc;
     thiscon->data = (char *) malloc(nchars+1);
